Rename hir_lowering to mir_lowering
